linux-and-64bit-cleanup-20050710
FIXES 19166

fix warnings found on x86_64 linux

+++ src/bozo/bnode.c
@@ -713,7 +713,7 @@ SendNotifierData(register int fd, register struct bnode_proc *tp)
     bufp += strlen(bufp);
     (void)sprintf(bufp, "END bnode\n");
     bufp += strlen(bufp);
-    len = (int)bufp - (int)buffer;
+    len = (int)(bufp - buffer);
     if (write(fd, buffer, len) < 0) {
 	return -1;
     }
